+++ src/extensions/domains/cover.extension.ts
@@ -79,7 +79,7 @@ export function VirtualCover({ context, synapse }: TServiceParams) {
     ],
   });
 
-  return <ATTRIBUTES extends object, LOCALS extends object>(
+  return <LOCALS extends object, ATTRIBUTES extends object>(
     options: AddEntityOptions<CoverConfiguration, CoverEvents, ATTRIBUTES, LOCALS>,
   ) => generate.addEntity(options);
 }
