+++ src/extensions/domains/date.extension.ts
@@ -31,7 +31,7 @@ export function VirtualDate({ context, synapse }: TServiceParams) {
     map_state: "native_value",
   });
 
-  return function <ATTRIBUTES extends object, LOCALS extends object>({
+  return function <LOCALS extends object, ATTRIBUTES extends object>({
     managed = true,
     ...options
   }: AddEntityOptions<DateConfiguration, DateEvents, ATTRIBUTES, LOCALS>) {
