+++ src/extensions/domains/number.extension.ts
@@ -52,7 +52,7 @@ export function VirtualNumber({ context, synapse }: TServiceParams) {
     ],
   });
 
-  return function <ATTRIBUTES extends object, LOCALS extends object>({
+  return function <LOCALS extends object, ATTRIBUTES extends object>({
     managed = true,
     ...options
   }: AddEntityOptions<NumberConfiguration, NumberEvents, ATTRIBUTES, LOCALS>) {
