+++ src/extensions/domains/remote.extension.ts
@@ -43,7 +43,7 @@ export function VirtualRemote({ context, synapse }: TServiceParams) {
     load_config_keys: ["current_activity", "activity_list", "supported_features"],
   });
 
-  return <ATTRIBUTES extends object, LOCALS extends object>(
+  return <LOCALS extends object, ATTRIBUTES extends object>(
     options: AddEntityOptions<RemoteConfiguration, RemoteEvents, ATTRIBUTES, LOCALS>,
   ) => generate.addEntity(options);
 }
