+++ src/extensions/domains/siren.extension.ts
@@ -35,7 +35,7 @@ export function VirtualSiren({ context, synapse }: TServiceParams) {
     load_config_keys: ["is_on", "available_tones", "supported_features"],
   });
 
-  return <ATTRIBUTES extends object, LOCALS extends object>(
+  return <LOCALS extends object, ATTRIBUTES extends object>(
     options: AddEntityOptions<SirenConfiguration, SirenEvents, ATTRIBUTES, LOCALS>,
   ) => generate.addEntity(options);
 }
