+++ src/extensions/domains/text.extension.ts
@@ -43,7 +43,7 @@ export function VirtualText({ context, synapse }: TServiceParams) {
     map_state: "native_value",
   });
 
-  return function <ATTRIBUTES extends object, LOCALS extends object>({
+  return function <LOCALS extends object, ATTRIBUTES extends object>({
     managed = true,
     ...options
   }: AddEntityOptions<TextConfiguration, TextEvents, ATTRIBUTES, LOCALS>) {
