+++ src/extensions/domains/time.extension.ts
@@ -27,7 +27,7 @@ export function VirtualTime({ context, synapse }: TServiceParams) {
     map_state: "native_value",
   });
 
-  return function <ATTRIBUTES extends object, LOCALS extends object>({
+  return function <LOCALS extends object, ATTRIBUTES extends object>({
     managed = true,
     ...options
   }: AddEntityOptions<TimeConfiguration, TimeEvents, ATTRIBUTES, LOCALS>) {
